requirements.txt
discord.py==2.7.1
requests==2.32.5
python-dotenv==1.2.2
